requirements.txt
Flask
face_recognition
Pillow
Werkzeug
Flask-SocketIO
eventlet 